Versão 1.0

+++ app/src/main/java/com/example/meutreino/ListarActivity.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import android.view.View
 import android.view.ViewGroup
 import android.widget.*
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_detalhe.*
 import kotlinx.android.synthetic.main.activity_detalhe.view.*
@@ -22,17 +23,23 @@ class ListarActivity: AppCompatActivity() {
 
     fun loadQuery(nomeTreino: String) {
         var database = Database(this)
-        val projections = arrayOf("_id", "circuito")
+        val projections = arrayOf("_id", "circuito", "atividade", "duracao", "data", "realizado")
         val selectionArgs = arrayOf(nomeTreino)
         val cursor = database.Select(projections, "circuito like ?", selectionArgs, "_id")
 
         listTreino.clear()
         if (cursor.moveToFirst()) {
             do {
-                val _id = cursor.getInt(cursor.getColumnIndex("_id"))
-                val circuito = cursor.getString(cursor.getColumnIndex("circuito"))
+                var treino: Treino = Treino()
+                treino._id = cursor.getInt(cursor.getColumnIndex("_id"))
+                treino.circuito = cursor.getString(cursor.getColumnIndex("circuito"))
+                treino.atividade = cursor.getString(cursor.getColumnIndex("atividade"))
+                treino.duracao = Integer.parseInt(cursor.getString(cursor.getColumnIndex("duracao")))
+                treino.data_treino = cursor.getString(cursor.getColumnIndex("data"))
+                val treinoRealizado = cursor.getInt(cursor.getColumnIndex("realizado")) > 0
+                treino.realizado = treinoRealizado
 
-                listTreino.add(Treino(_id, circuito))
+                listTreino.add(treino)
             } while (cursor.moveToNext())
         }
 
@@ -57,38 +64,47 @@ class ListarActivity: AppCompatActivity() {
 
             viewDetalhe.tvId.text = treinoAtual._id.toString()
             viewDetalhe.tvCircuito.text = treinoAtual.circuito
+            viewDetalhe.tvAtividade.text = treinoAtual.atividade
+            viewDetalhe.tvDuracao.text = treinoAtual.duracao.toString() + " minutos"
 
-            viewDetalhe.ivSave.setOnClickListener(View.OnClickListener {
-                var database = Database(this.context!!)
-                val selectionArgs = arrayOf(treinoAtual._id.toString())
-
-                Toast.makeText(this.context, " Editando! ", Toast.LENGTH_SHORT).show()
-                viewDetalhe.ivEdit.visibility = View.VISIBLE
-                viewDetalhe.ivDelete.visibility = View.VISIBLE
-            })
+            val arrData = treinoAtual.data_treino.split('-')
+            val dataFormatada = arrData[2].padStart(2, '0') + "/" + arrData[1] + "/" + arrData[0]
+            viewDetalhe.tvData.text = dataFormatada
 
             viewDetalhe.ivEdit.setOnClickListener(View.OnClickListener {
                 var database = Database(this.context!!)
                 val selectionArgs = arrayOf(treinoAtual._id.toString())
 
                 Toast.makeText(this.context, " Editando! ", Toast.LENGTH_SHORT).show()
-                viewDetalhe.ivEdit.visibility = View.GONE
-                viewDetalhe.ivDelete.visibility = View.GONE
 
                 val intent = Intent(this.context, CadastrarActivity::class.java)
-                intent.putExtra("circuito", listTreino.get(position).circuito)
                 intent.putExtra("_id", listTreino.get(position)._id)
+                intent.putExtra("circuito", listTreino.get(position).circuito)
+                intent.putExtra("atividade", listTreino.get(position).atividade)
+                intent.putExtra("duracao", listTreino.get(position).duracao)
+                intent.putExtra("data", listTreino.get(position).data_treino)
+                intent.putExtra("realizado", listTreino.get(position).realizado)
                 startActivity(intent)
             })
 
             viewDetalhe.ivDelete.setOnClickListener(View.OnClickListener {
-                var database = Database(this.context!!)
-                val selectionArgs = arrayOf(treinoAtual._id.toString())
+                val builder = AlertDialog.Builder(this@ListarActivity)
+                builder.setMessage("Tem certeza que deseja excluir?")
+                    .setPositiveButton("Sim") { dialog, id ->
+                        var database = Database(this.context!!)
+                        val selectionArgs = arrayOf(treinoAtual._id.toString())
 
-                Toast.makeText(this.context, " Deletando! ", Toast.LENGTH_SHORT).show()
-                database.Delete("_id=?", selectionArgs)
-                viewDetalhe.visibility = View.GONE
-                loadQuery("%")
+                        Toast.makeText(this.context, " Excluído! ", Toast.LENGTH_SHORT).show()
+                        database.Delete("_id=?", selectionArgs)
+                        viewDetalhe.visibility = View.GONE
+                        loadQuery("%")
+                    }
+                    .setNegativeButton("Não") { dialog, id ->
+                        dialog.dismiss()
+                    }
+                val alert = builder.create()
+                alert.show()
+
             })
 
             return viewDetalhe;
